update summary script read_in_files function
in progress of fixing issue where data weren't able to be read in because the pivot longer wasn't working when there are data columns that aren't all numeric.
next steps: fix UI for asking about text in values. also need to figure out why R is reading in 0.00007 in as 7e-04 for the Sediment_Incubations_Respiration_Rates file.