change checks report from xlsx to csv
It was previously xlsx because that format caused fewer issues when opening it in excel. however, with large data packages (jake's) it was failing and preventing the report from generating. I changed it to csv so now it works with larger DPs (but you will have to do some format fixing in excel if you wanna look at the table)