Update load_tabular_data_from_flmd.R
Finished refactoring read_csv/tsv for when query_header_info == F. Now, when there are no header info, it reads in with read_csv()/read_tsv() with most of those functions default arguments. I left query_header_info == T as it was (which reads top few lines, asked user for the row the col names are, then again for the first line of the data). then it uses read_csv/tsv twice; first to read in headers, second to read in data. then it smooshes cols names onto data. query_header_info is a lot slower due to having to read in the data multiple times