Update data_package_checks.R
Addressed Brie's comments about changing code to better handle when an FLMD is present. I added a user input to ask if there's an FLMD present and if there is for the file path. And then removed the create_flmd_skeleton() from this code. It will now either use an already existing FLMD that has header info. Or will rely on the built in functionality of the load_tabular_data_from_flmd() function to grab the info it needs

Co-Authored-By: Brieanne Forbes <108300528+brieanne-forbes@users.noreply.github.com>