Fix create_flmd() header row info calcs
Code was previously carried over when I was reading in data where col_names = T. Now that col_names = F, the numbers were messed up. Also, there was a line that was calculating header rows based on a previous (incorrect) method. This commit fixed that and aligns these calculations to the recent messages from Dylan