fix create_flmd calculations
There was a bug in the loop that was asking for header format standards. For example, if a user picked "other" initially it would ask for the header info. Then ask if a standard should be applied. If the user selected "n" for no/none, then on the next loop, it would see the (new) "n" and overwrite that header info as the default 1,1 value.